refactor: fold legacy chat modules into chat
