+++ src/gigachat/api/__init__.py
@@ -5,7 +5,6 @@ from gigachat.api import (
     chat_completions,
     embeddings,
     files,
-    legacy_chat,
     models,
     threads,
     tools,
@@ -19,7 +18,6 @@ __all__ = [
     "chat_completions",
     "embeddings",
     "files",
-    "legacy_chat",
     "models",
     "threads",
     "tools",
