+++ src/gigachat/models/chat.py
@@ -1,21 +1,231 @@
-from gigachat.models.legacy_chat import (
-    LegacyChat,
-    LegacyChatCompletion,
-    LegacyChatCompletionChunk,
-    LegacyChatFunctionCall,
-    LegacyChoice,
-    LegacyChoiceChunk,
-    LegacyFewShotExample,
-    LegacyFunction,
-    LegacyFunctionCall,
-    LegacyFunctionParameters,
-    LegacyFunctionParametersProperty,
-    LegacyMessage,
-    LegacyMessageChunk,
-    LegacyMessageRole,
-    LegacyStorage,
-    LegacyUsage,
-)
+from enum import Enum
+from typing import Any, Dict, List, Literal, Optional, Union
+
+from pydantic import BaseModel, ConfigDict, Field, model_validator
+
+from gigachat.models.base import APIResponse
+from gigachat.models.response_format import ResponseFormat
+
+
+class LegacyMessageRole(str, Enum):
+    """Role of the message author."""
+
+    ASSISTANT = "assistant"
+    SYSTEM = "system"
+    USER = "user"
+    FUNCTION = "function"
+    SEARCH_RESULT = "search_result"
+    FUNCTION_IN_PROGRESS = "function_in_progress"
+
+
+class LegacyFunctionCall(BaseModel):
+    """Model function call."""
+
+    name: str = Field(description="Name of the function to call.")
+    arguments: Optional[Dict[Any, Any]] = Field(default=None, description="Function call arguments.")
+
+
+class LegacyFewShotExample(BaseModel):
+    """Few-shot example for function definition."""
+
+    request: str = Field(description="User request.")
+    params: Dict[str, Any] = Field(description="Parameters corresponding to the request.")
+
+
+class LegacyStorage(BaseModel):
+    """Context storage settings for GigaChat."""
+
+    is_stateful: bool = Field(
+        description=(
+            "Context storage mode. If True, context messages don't need to be passed every time. "
+            "Only the new message is required."
+        )
+    )
+    limit: Optional[int] = Field(
+        default=None,
+        description=(
+            "Maximum number of historical context messages sent to the model. "
+            "If the user passes a limit and has a system prompt or assistant instruction, "
+            "the number of messages sent is limit + 1 (adding the system prompt). "
+            "If not provided, the entire context is sent."
+        ),
+    )
+    assistant_id: Optional[str] = Field(
+        default=None,
+        description=(
+            "Assistant ID when creating a thread (only in the first message). "
+            "If provided, 'model' cannot be provided. If thread_id is provided, this is not passed."
+        ),
+    )
+    thread_id: Optional[str] = Field(default=None, description="Thread ID. Not populated for the first message.")
+    metadata: Optional[Dict[Any, Any]] = Field(default=None, description="Arbitrary metadata.")
+
+
+class LegacyUsage(BaseModel):
+    """Model usage statistics."""
+
+    prompt_tokens: int = Field(description="Number of tokens in the input message.")
+    completion_tokens: int = Field(description="Number of tokens generated by the model.")
+    total_tokens: int = Field(description="Total number of tokens.")
+    precached_prompt_tokens: Optional[int] = Field(default=None, description="Number of tokens served from cache.")
+
+
+class LegacyFunctionParametersProperty(BaseModel):
+    """Property of a function parameter."""
+
+    type_: str = Field(default="object", alias="type", description="Type of the argument.")
+    description: str = Field(default="", description="Description of the argument.")
+    items: Optional[Dict[str, Any]] = Field(default=None, description="Items schema for array types.")
+    enum: Optional[List[str]] = Field(default=None, description="List of possible values for enum types.")
+    properties: Optional[Dict[Any, "LegacyFunctionParametersProperty"]] = Field(
+        default=None, description="Nested properties for object types."
+    )
+
+
+class LegacyFunctionParameters(BaseModel):
+    """Parameters definition for a function."""
+
+    type_: str = Field(default="object", alias="type", description="Type of the parameters object (usually 'object').")
+    properties: Optional[Dict[Any, LegacyFunctionParametersProperty]] = Field(
+        default=None, description="Dictionary of parameter properties."
+    )
+    required: Optional[List[str]] = Field(default=None, description="List of required parameter names.")
+
+
+class LegacyFunction(BaseModel):
+    """Function definition that can be called by the model."""
+
+    name: str = Field(description="Name of the function.")
+    description: Optional[str] = Field(default=None, description="Description of the function.")
+    parameters: Optional[LegacyFunctionParameters] = Field(default=None, description="Parameters definition.")
+    few_shot_examples: Optional[List[LegacyFewShotExample]] = Field(
+        default=None, description="List of few-shot examples for the function."
+    )
+    return_parameters: Optional[Dict[Any, Any]] = Field(default=None, description="Description of return parameters.")
+
+    @model_validator(mode="before")
+    @classmethod
+    def _fix_title_and_parameters(cls, values: Any) -> Any:
+        """Pydantic adapter (title -> name), (parameters -> properties)."""
+        if isinstance(values, dict):
+            values = dict(values)
+
+            if values.get("name") in (None, "") and values.get("title"):
+                values["name"] = values.pop("title", None)
+
+            if values.get("parameters") in (None, "", {}) and "properties" in values:
+                values["parameters"] = {
+                    "properties": values.pop("properties", {}),
+                }
+
+        return values
+
+
+class LegacyChatFunctionCall(BaseModel):
+    """Specific function call request."""
+
+    name: str = Field(description="Name of the function.")
+    partial_arguments: Optional[Dict[str, Any]] = Field(default=None, description="Partial arguments for the function.")
+
+
+class LegacyMessage(BaseModel):
+    """Message in a chat conversation."""
+
+    role: LegacyMessageRole = Field(description="Role of the message author.")
+    content: str = Field(default="", description="Text content of the message.")
+    function_call: Optional[LegacyFunctionCall] = Field(default=None, description="Function call details.")
+    name: Optional[str] = Field(default=None, description="Function name. Required if role is 'function'.")
+    attachments: Optional[List[str]] = Field(default=None, description="List of attached file IDs.")
+    data_for_context: Optional[List["LegacyMessage"]] = Field(default=None, description="DEPRECATED: Data for context.")
+    functions_state_id: Optional[str] = Field(
+        default=None, description="ID of the function state generating images/video."
+    )
+    reasoning_content: Optional[str] = Field(default=None, description="Reasoning content from the model.")
+    id_: Optional[Any] = Field(alias="id", default=None, description="Message ID.")
+
+    model_config = ConfigDict(use_enum_values=True)
+
+
+class LegacyMessageChunk(BaseModel):
+    """Chunk of a message in a stream."""
+
+    role: Optional[LegacyMessageRole] = Field(default=None, description="Role of the message author.")
+    content: Optional[str] = Field(default=None, description="Text content chunk.")
+    reasoning_content: Optional[str] = Field(default=None, description="Reasoning content chunk.")
+    function_call: Optional[LegacyFunctionCall] = Field(default=None, description="Function call chunk.")
+    functions_state_id: Optional[str] = Field(default=None, description="Function state ID.")
+
+
+class LegacyChoice(BaseModel):
+    """Completion choice."""
+
+    message: LegacyMessage = Field(description="Generated message.")
+    index: int = Field(description="Index of the choice in the list.")
+    finish_reason: Optional[str] = Field(default=None, description="Reason why the generation finished.")
+
+
+class LegacyChoiceChunk(BaseModel):
+    """Completion choice chunk in a stream."""
+
+    delta: LegacyMessageChunk = Field(description="Message delta.")
+    index: int = Field(description="Index of the choice in the list.")
+    finish_reason: Optional[str] = Field(default=None, description="Reason why the generation finished.")
+
+
+class LegacyChat(BaseModel):
+    """Chat completion request parameters."""
+
+    model: Optional[str] = Field(default=None, description="Name of the model to use.")
+    messages: List[LegacyMessage] = Field(description="List of messages in the conversation.")
+    temperature: Optional[float] = Field(default=None, description="Sampling temperature.")
+    top_p: Optional[float] = Field(default=None, description="Nucleus sampling parameter (alternative to temperature).")
+    n: Optional[int] = Field(default=None, description="Number of completion choices to generate.")
+    stream: Optional[bool] = Field(default=None, description="If True, stream partial progress.")
+    max_tokens: Optional[int] = Field(default=None, description="Maximum number of tokens to generate.")
+    repetition_penalty: Optional[float] = Field(default=None, description="Repetition penalty factor.")
+    update_interval: Optional[float] = Field(default=None, description="Interval in seconds between stream updates.")
+    profanity_check: Optional[bool] = Field(default=None, description="Enable profanity filtering.")
+    function_call: Optional[Union[Literal["auto", "none"], LegacyChatFunctionCall]] = Field(
+        default=None, description="Controls which function is called."
+    )
+    functions: Optional[List[LegacyFunction]] = Field(
+        default=None, description="List of functions available to the model."
+    )
+    flags: Optional[List[str]] = Field(default=None, description="List of feature flags.")
+    storage: Optional[LegacyStorage] = Field(default=None, description="Context storage settings.")
+    response_format: Optional[ResponseFormat] = Field(default=None, description="Response format (e.g. JSON Schema).")
+    additional_fields: Optional[Dict[str, Any]] = Field(
+        default=None, description="Additional fields to pass to the API."
+    )
+    reasoning_effort: Optional[Literal["low", "medium", "high"]] = Field(
+        default=None, description="Reasoning effort level."
+    )
+
+
+class LegacyChatCompletion(APIResponse):
+    """Chat completion response."""
+
+    choices: List[LegacyChoice] = Field(description="List of completion choices.")
+    created: int = Field(description="Creation timestamp (Unix time).")
+    model: str = Field(description="Model name used for generation.")
+    thread_id: Optional[str] = Field(default=None, description="Thread ID.")
+    message_id: Optional[str] = Field(default=None, description="Message ID. Present if storage mode is used.")
+    usage: LegacyUsage = Field(description="Usage statistics.")
+    object_: str = Field(alias="object", description="Object type (e.g. 'chat.completion').")
+
+
+class LegacyChatCompletionChunk(APIResponse):
+    """Chat completion response chunk."""
+
+    choices: List[LegacyChoiceChunk] = Field(description="List of completion choice chunks.")
+    created: int = Field(description="Creation timestamp (Unix time).")
+    model: str = Field(description="Model name used for generation.")
+    object_: str = Field(alias="object", description="Object type (e.g. 'chat.completion.chunk').")
+    usage: Optional[LegacyUsage] = Field(default=None, description="Usage statistics.")
+
+
+LegacyFunctionParametersProperty.model_rebuild()
+LegacyMessage.model_rebuild()
 
 
 class FunctionCall(LegacyFunctionCall):
